moved class to appropriate package and added license header
