+++ src/main/java/it/smartcommunitylab/aac/saml/auth/SamlWebSsoAuthenticationFilter.java
@@ -23,7 +23,7 @@ import it.smartcommunitylab.aac.core.auth.UserAuthentication;
 import it.smartcommunitylab.aac.core.auth.WebAuthenticationDetails;
 import it.smartcommunitylab.aac.core.provider.ProviderConfigRepository;
 import it.smartcommunitylab.aac.saml.SamlIdentityAuthority;
-import it.smartcommunitylab.aac.saml.provider.Saml2AuthenticationTokenConverter;
+import it.smartcommunitylab.aac.saml.service.Saml2AuthenticationTokenConverter;
 import it.smartcommunitylab.aac.saml.provider.SamlIdentityProviderConfig;
 import it.smartcommunitylab.aac.saml.service.HttpSessionSaml2AuthenticationRequestRepository;
 import java.io.IOException;
