+++ src/main/java/it/smartcommunitylab/aac/saml/service/Saml2AuthenticationTokenConverter.java
@@ -1,9 +1,22 @@
-// TODO: file header according to AAC guidelines
+/*
+ * Copyright 2023 the original author or authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
-package it.smartcommunitylab.aac.saml.provider;
+package it.smartcommunitylab.aac.saml.service;
 
 import it.smartcommunitylab.aac.saml.auth.SerializableSaml2AuthenticationRequestContext;
-import it.smartcommunitylab.aac.saml.service.HttpSessionSaml2AuthenticationRequestRepository;
 
 import java.io.ByteArrayOutputStream;
 import java.nio.charset.StandardCharsets;
